api/requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
pydantic
sentence-transformers
chromadb
beautifulsoup4
requests
apscheduler
